@@ -15,52 +15,53 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.0.0",
-    "@angular/common": "~7.0.0",
-    "@angular/compiler": "~7.0.0",
-    "@angular/core": "~7.0.0",
-    "@angular/forms": "~7.0.0",
-    "@angular/http": "~7.0.0",
-    "@angular/platform-browser": "~7.0.0",
-    "@angular/platform-browser-dynamic": "~7.0.0",
-    "@angular/router": "~7.0.0",
+    "@angular/animations": "~8.2.14",
+    "@angular/common": "~8.2.14",
+    "@angular/compiler": "~8.2.14",
+    "@angular/core": "~8.2.14",
+    "@angular/forms": "~8.2.14",
+    "@angular/platform-browser": "~8.2.14",
+    "@angular/platform-browser-dynamic": "~8.2.14",
+    "@angular/router": "~8.2.14",
     "@types/jquery": "^3.3.31",
     "angular-cli-ghpages": "^0.6.0",
-    "core-js": "^2.5.4",
-    "echarts": "^4.2.1",
+    "core-js": "^3.6.0",
+    "echarts": "^4.5.0",
     "jquery": "^3.4.1",
     "loadsh": "0.0.4",
-    "ngx-echarts": "^4.1.0",
-    "node-sass": "^4.12.0",
-    "rxjs": "~6.3.3",
-    "rxjs-compat": "^6.4.0",
-    "zone.js": "~0.8.26"
+    "ngx-echarts": "^4.2.2",
+    "node-sass": "^4.13.0",
+    "rxjs": "~6.5.3",
+    "rxjs-compat": "^6.5.3",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.12.4",
-    "@angular/cli": "^7.0.1",
-    "@angular/compiler-cli": "^7.2.14",
-    "@angular/language-service": "~7.0.0",
-    "@types/echarts": "^4.1.8",
-    "@types/jasmine": "^2.6.3",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/lodash": "^4.14.123",
-    "@types/node": "~8.9.4",
+    "@angular-devkit/build-angular": "~0.803.21",
+    "@angular/cli": "^8.3.21",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "~8.2.14",
+    "@types/echarts": "^4.4.2",
+    "@types/jasmine": "^3.5.0",
+    "@types/jasminewd2": "~2.0.8",
+    "@types/lodash": "^4.14.149",
+    "@types/node": "~12.12.21",
     "angular-fontawesome": "^1.0.0",
     "angular-router-loader": "^0.8.5",
-    "angular-tree-component": "^8.4.0",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
+    "angular-tree-component": "^8.5.2",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
     "jquery": "^3.4.1",
-    "karma": "~3.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma": "~4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
     "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.1.1"
+    "raw-loader": "^1.0.0",
+    "ts-node": "~8.5.4",
+    "tslint": "~5.20.1",
+    "typescript": "^3.4.3"
   }
 }
